refactor(uimodel): 패키지 이동
