+++ app/src/main/java/woowacourse/shopping/mapper/ProductMapper.kt
@@ -1,7 +1,7 @@
-package woowacourse.shopping.ui.mapper
+package woowacourse.shopping.mapper
 
 import woowacourse.shopping.domain.Product
-import woowacourse.shopping.ui.model.UiProduct
+import woowacourse.shopping.model.UiProduct
 
 fun UiProduct.toDomain(): Product =
     Product(id = id, name = name, price = price.toDomain(), imageUrl = imageUrl)
