+++ app/src/main/java/woowacourse/shopping/mapper/RecentProductMapper.kt
@@ -1,7 +1,7 @@
-package woowacourse.shopping.ui.mapper
+package woowacourse.shopping.mapper
 
 import woowacourse.shopping.domain.RecentProduct
-import woowacourse.shopping.ui.model.UiRecentProduct
+import woowacourse.shopping.model.UiRecentProduct
 
 fun UiRecentProduct.toDomain(): RecentProduct =
     RecentProduct(id = id, product = product.toDomain())
