+++ app/src/main/java/woowacourse/shopping/model/Price.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.model
+package woowacourse.shopping.model
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
