+++ app/src/main/java/woowacourse/shopping/model/RecentProduct.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.model
+package woowacourse.shopping.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
