+++ app/src/main/java/woowacourse/shopping/ui/shopping/recyclerview/product/ProductViewHolder.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ItemProductBinding
-import woowacourse.shopping.ui.model.UiProduct
+import woowacourse.shopping.model.UiProduct
 import woowacourse.shopping.util.setOnSingleClickListener
 
 class ProductViewHolder(parent: ViewGroup, onItemClick: (Int) -> Unit) :
