+++ app/src/main/res/layout/item_recent_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="recentProduct"
-            type="woowacourse.shopping.ui.model.RecentProduct" />
+            type="woowacourse.shopping.model.RecentProduct" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
